refactor(app): #31 moved tool management components to modules
